python/requirements.txt
python-dotenv
minio
waitress